requirements.txt
alpaca-py>=0.35.0
apscheduler>=3.10.0
hmmlearn>=0.3.3
joblib>=1.4.0
nbformat>=4.2.0
numpy>=2.0.0
polars>=1.20.0
pyarrow>=19.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytz>=2024.1
pyyaml>=6.0.2
scikit-learn>=1.5.0
tqdm>=4.66.0
yfinance>=0.2.54